+++ acquire-core-libs/src/acquire-device-properties/device/props/storage.h
@@ -42,8 +42,11 @@ extern "C"
     /// Properties for a storage driver.
     struct StorageProperties
     {
-        struct String filename;
+        struct String uri;
         struct String external_metadata_json;
+        struct String access_key_id;
+        struct String secret_access_key;
+
         uint32_t first_frame_id;
         struct PixelScale pixel_scale_um;
 
@@ -68,6 +71,7 @@ extern "C"
         uint8_t chunking_is_supported;
         uint8_t sharding_is_supported;
         uint8_t multiscale_is_supported;
+        uint8_t s3_is_supported;
     };
 
     /// Initializes StorageProperties, allocating string storage on the heap
@@ -76,10 +80,10 @@ extern "C"
     /// @param[out] out The constructed StorageProperties object.
     /// @param[in] first_frame_id (unused; aiming for future file rollover
     /// support
-    /// @param[in] filename A c-style null-terminated string. The file to create
-    ///                     for streaming.
-    /// @param[in] bytes_of_filename Number of bytes in the `filename` buffer
-    ///                              including the terminating null.
+    /// @param[in] uri A c-style null-terminated string. The file to create
+    ///                for streaming.
+    /// @param[in] bytes_of_uri Number of bytes in the `uri` buffer,
+    ///                         including the terminating null.
     /// @param[in] metadata A c-style null-terminated string. Metadata string
     ///                     to save along side the created file.
     /// @param[in] bytes_of_metadata Number of bytes in the `metadata` buffer
@@ -90,8 +94,8 @@ extern "C"
     /// to zero.
     int storage_properties_init(struct StorageProperties* out,
                                 uint32_t first_frame_id,
-                                const char* filename,
-                                size_t bytes_of_filename,
+                                const char* uri,
+                                size_t bytes_of_uri,
                                 const char* metadata,
                                 size_t bytes_of_metadata,
                                 struct PixelScale pixel_scale_um,
@@ -105,27 +109,46 @@ extern "C"
     int storage_properties_copy(struct StorageProperties* dst,
                                 const struct StorageProperties* src);
 
-    /// @brief Set the filename string in `out`.
+    /// @brief Set the uri string in `out`.
     /// Copies the string into storage owned by the properties struct.
     /// @returns 1 on success, otherwise 0
     /// @param[in,out] out The storage properties to change.
-    /// @param[in] filename pointer to the beginning of the filename buffer.
-    /// @param[in] bytes_of_filename the number of bytes in the filename buffer.
-    ///                              Should include the terminating NULL.
-    int storage_properties_set_filename(struct StorageProperties* out,
-                                        const char* filename,
-                                        size_t bytes_of_filename);
+    /// @param[in] uri Pointer to the beginning of the uri buffer.
+    /// @param[in] bytes_of_uri The number of bytes in the uri buffer.
+    ///                         Should include the terminating NULL.
+    int storage_properties_set_uri(struct StorageProperties* out,
+                                   const char* uri,
+                                   size_t bytes_of_uri);
 
     /// @brief Set the metadata string in `out`.
     /// Copies the string into storage owned by the properties struct.
     /// @returns 1 on success, otherwise 0
     /// @param[in,out] out The storage properties to change.
     /// @param[in] metadata pointer to the beginning of the metadata buffer.
-    /// @param[in] bytes_of_filename the number of bytes in the metadata buffer.
+    /// @param[in] bytes_of_metadata the number of bytes in the metadata buffer.
     ///                              Should include the terminating NULL.
     int storage_properties_set_external_metadata(struct StorageProperties* out,
                                                  const char* metadata,
                                                  size_t bytes_of_metadata);
+
+    /// @brief Set the access key id string in `out`.
+    /// Copies the string into storage owned by the properties struct.
+    /// @returns 1 on success, otherwise 0
+    /// @param[in,out] out The storage properties to change.
+    /// @param[in] access_key_id Pointer to the beginning of the access key id
+    ///                          buffer.
+    /// @param[in] bytes_of_access_key_id The number of bytes in the access key
+    ///                                   id.
+    /// @param[in] secret_access_key Pointer to the beginning of the secret
+    ///                              access key buffer.
+    /// @param[in] bytes_of_secret_access_key The number of bytes in the secret
+    ///                                       access key.
+    int storage_properties_set_access_key_and_secret(
+      struct StorageProperties* out,
+      const char* access_key_id,
+      size_t bytes_of_access_key_id,
+      const char* secret_access_key,
+      size_t bytes_of_secret_access_key);
 
     /// @brief Set the value of the StorageDimension struct at index `index` in
     /// `out`.